services/ocr/requirements.txt
flask>=3.0
pytesseract>=0.3.10
easyocr>=1.7
opencv-python-headless>=4.9
pillow>=10.2
numpy>=1.26
spacy>=3.7
pdf2image>=1.17
PyMuPDF>=1.23
python-magic>=0.4.27
editdistance>=0.8
gunicorn>=21.2
pytest>=8.0
openai>=1.0
scipy>=1.11
